+++ pkg/azuremachinepool/mutate_azuremachinepool_create.go
@@ -8,22 +8,34 @@ import (
 	"github.com/giantswarm/micrologger"
 	"k8s.io/api/admission/v1beta1"
 	expcapzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1alpha3"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/azure-admission-controller/internal/vmcapabilities"
+	"github.com/giantswarm/azure-admission-controller/pkg/generic"
 	"github.com/giantswarm/azure-admission-controller/pkg/mutator"
 )
 
 type CreateMutator struct {
-	logger micrologger.Logger
-	vmcaps *vmcapabilities.VMSKU
+	ctrlClient client.Client
+	location   string
+	logger     micrologger.Logger
+	vmcaps     *vmcapabilities.VMSKU
 }
 
 type CreateMutatorConfig struct {
-	Logger micrologger.Logger
-	VMcaps *vmcapabilities.VMSKU
+	CtrlClient client.Client
+	Location   string
+	Logger     micrologger.Logger
+	VMcaps     *vmcapabilities.VMSKU
 }
 
 func NewCreateMutator(config CreateMutatorConfig) (*CreateMutator, error) {
+	if config.CtrlClient == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.CtrlClient must not be empty", config)
+	}
+	if config.Location == "" {
+		return nil, microerror.Maskf(invalidConfigError, "%T.Location must not be empty", config)
+	}
 	if config.Logger == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
 	}
@@ -32,8 +44,10 @@ func NewCreateMutator(config CreateMutatorConfig) (*CreateMutator, error) {
 	}
 
 	m := &CreateMutator{
-		logger: config.Logger,
-		vmcaps: config.VMcaps,
+		ctrlClient: config.CtrlClient,
+		location:   config.Location,
+		logger:     config.Logger,
+		vmcaps:     config.VMcaps,
 	}
 
 	return m, nil
@@ -52,7 +66,15 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 		return []mutator.PatchOperation{}, microerror.Maskf(parsingFailedError, "unable to parse azureMachinePool CR: %v", err)
 	}
 
-	patch, err := m.ensureStorageAccountType(ctx, azureMPCR)
+	patch, err := m.ensureLocation(ctx, azureMPCR)
+	if err != nil {
+		return []mutator.PatchOperation{}, microerror.Mask(err)
+	}
+	if patch != nil {
+		result = append(result, *patch)
+	}
+
+	patch, err = m.ensureStorageAccountType(ctx, azureMPCR)
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}
@@ -61,6 +83,14 @@ func (m *CreateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 	}
 
 	patch, err = m.ensureDataDisks(ctx, azureMPCR)
+	if err != nil {
+		return []mutator.PatchOperation{}, microerror.Mask(err)
+	}
+	if patch != nil {
+		result = append(result, *patch)
+	}
+
+	patch, err = generic.EnsureReleaseVersionLabel(ctx, m.ctrlClient, azureMPCR.GetObjectMeta())
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}
@@ -83,8 +113,15 @@ func (m *CreateMutator) ensureStorageAccountType(ctx context.Context, mpCR *expc
 	if mpCR.Spec.Template.OSDisk.ManagedDisk.StorageAccountType == "" {
 		// We need to set the default value as it is missing.
 
+		location := mpCR.Spec.Location
+		if location == "" {
+			// The location was empty and we are adding it using this same mutator.
+			// We assume it will be set to the installation's location.
+			location = m.location
+		}
+
 		// Check if the VM has Premium Storage capability.
-		premium, err := m.vmcaps.HasCapability(ctx, mpCR.Spec.Location, mpCR.Spec.Template.VMSize, vmcapabilities.CapabilityPremiumIO)
+		premium, err := m.vmcaps.HasCapability(ctx, location, mpCR.Spec.Template.VMSize, vmcapabilities.CapabilityPremiumIO)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -110,4 +147,12 @@ func (m *CreateMutator) ensureDataDisks(ctx context.Context, mpCR *expcapzv1alph
 	}
 
 	return mutator.PatchAdd("/spec/template/dataDisks", desiredDataDisks), nil
+}
+
+func (m *CreateMutator) ensureLocation(ctx context.Context, mpCR *expcapzv1alpha3.AzureMachinePool) (*mutator.PatchOperation, error) {
+	if len(mpCR.Spec.Location) > 0 {
+		return nil, nil
+	}
+
+	return mutator.PatchAdd("/spec/location", m.location), nil
 }